requirements.txt
email-validator~=2.0
socid-extractor~=0.0.26
socialshares~=1.1.0
TikTokApi~=6.1.1
fastapi~=0.104.1
html5lib~=1.1
selenium~=4.15.2
python-dotenv~=1.0.0
tomd~=0.1.3
pandoc~=2.3
pytrends~=4.9.2
tweepy~=4.14.0
python-decouple~=3.8
pypandoc~=1.12
prompt_toolkit~=3.0.41
rich~=13.6.0
uvicorn~=0.24.0.post1
pydantic~=2.5.1
google-search-results~=2.4.2
socials~=0.2.0
feedfinder2~=0.0.4
pika~=1.3.2
pyseoanalyzer~=4.0.7
markdownify~=0.11.6
newspaper3k~=0.2.8
feedparser~=6.0.10
pydantic~=2.5.1
pysocks~=1.7.1
requests~=2.27.0
GoogleNews~=1.6.10
vaderSentiment~=3.3.2
spacy~=3.7.2
spacy-html-tokenizer~=0.1.3
spacy-dbpedia-spotlight~=0.2.6
nltk~=3.8.1
textblob~=0.17.1
youtube-search~=2.1.2
cachetools~=5.3.2



